Make CherryPick lazily demand the CurrentUser
The CherryPick view is created to test if it implements the UiCommand
interface, but it doesn't need to create the CherryPickChange
implementation until apply().  Deferring creation with a Provider also
defers attempting to cast the CurrentUser into an IdentifiedUser, and
fixes an exception thrown while inspecting the possible views of a
revision inside of PatchSetDetailFactory.

Change-Id: I13968308694664c760f1228db9fabd482850b62e

Former-commit-id: 4548625e6f1244257711fd1c9e1dee4cd2bb1433