Document the 'Buck Exit Criteria'
This came from the repo-discuss mailing list thread[1] and is a
summary of the major criteria that has been proposed. As issues
are resolved we will delete these entries from the documentation,
until either the section itself can be removed, or we have reached
September 1st 2013 (~3 months) and the build issue can be revisited.

[1] https://groups.google.com/d/msg/repo-discuss/gt2Oeb9LkeI/xsX5HKMH5QoJ

Change-Id: Id9b7dacb75fe8bd74d45e9ed141ac4799314a331

Former-commit-id: cb486dd013897763bb9d5ff718ca54a1d01a1b56