Expand the dependencies section in the buck documentation
Mention that it is currently only possible to download dependencies
over HTTP from the source sites, and `http_proxy` should be set when
running behind an HTTP proxy.

Also make sure we consistently add two blank lines before headlines,
and one blank line after, in the rest of the document.

Change-Id: I4c85fe5f47a9d78b6c07eee18f2173980ce5c25f

Former-commit-id: b46cf494fd6ceb5d3168fc1224b35af9960047ab