Update the buck installation documentation
Instead of adding the buck build's output directory in the path, we
make a symbolic link to the executable from `~/bin`.

Instrutions for setting up ~/bin are copied from the repo installation
guide at Android [1].

The sections for Linux and Mac OS are merged to a single section, and
the summary now mentions that only Linux and Mac OS are supported.

[1] http://source.android.com/source/downloading.html

Change-Id: Iaa973ec4116b03371ea386e4286c8bbfa36e7190

Former-commit-id: 64c1644df58c8242f3dbe1ec3a2841bc4641e9bb