Use --proxy-anyauth when running curl
This option should permit curl to pick up proxy authentication
information from ~/.netrc rather than requiring it to be embedded into
the http_proxy environment variable.  This is important on shared user
systems where the environment of a process is typically readable by
any user, while ~/.netrc is usually set with a mode 0600.

Change-Id: I995e6c8e9677dbda14a4e79942eed6682e381c36

Former-commit-id: 7e2eff6e4d6d15f053e84178900b11fc799251fe