Fix download target to only download binary JARs
Previously it also setup the prebuilt_jar() which isn't actually
necessary to execute the download. This reduces the size of the
command line passed into the child buck process.

Before this change not all libraries were downloaded during the
download target. Targets in the root of //lib were skipped due to
an incorrect regex passed to egrep. This has been fixed with the
new regex inside of download_all.py.

The initial binary download now requires ~59s on my home cable modem,
and pulls 84M.

Source JARs are now only downloaded when the user builds the
download_sources target.  This saves ~30s of setup time on a
the same connection and reduces the initial download by 16M.

Change-Id: Ie3bc97c9101c94f8bc4b89664bbd218323ad7230

Former-commit-id: 7b55326512273825a4a86989fd6fc2bd0f0e0ccf