Require preferred email to be a verified address
Warn the user if they attempt to select a preferred email
address that has not been previously verified and stored in the
account_external_ids table. This reduces the chances the server
will have email notifications bounce because a user subscribed to
changes and entered an invalid preferred email address.

Change-Id: Ib00fc25fb11445968bc4114a2b16ccb08437a699

Former-commit-id: 4f6c76e758b3f1b279e720c285332db76cc03abb