Fix PatchScript's mapping of lines below the last edit
For comments below the last edit block of a PatchScript, the formulae
to obtain the line number shifts between sides subtracted one side's
end, but added the other side's beginning (instead of also the
end). This mismatch could cause the client to fetch lines beyond a
file's end and thereby cause an ArrayIndexofBoundException.

We now use the difference of the last edit block's ends to map lines
below the last edit.

Change-Id: I67d2baa0f5ede3af7348c609034de81e8acf9c17

Former-commit-id: 490fa71cc798c87e0d0aaace754584c6d702cd83