Set python version back to 2.6.5 in .pydevproject
When I fixed some unused imports in 0c0040b I did `git commit -a`
and pushed without realising that Eclipse had changed the Python
version in the PyDev project config.

This change sets the version back to 2.6.5 which is the minimum
version required to run the python scripts in the Buck build.

Change-Id: Idc97af3d01999889d9b1a818fbd1bbe0b274dcf3

Former-commit-id: 77e974c7070e274aaca3f2413a3fb53031d0f50e