Minor cleanups in Emails collection
The email cache is not necessary here. The parent resource has
an instance of IdentifiedUser, which contains the verified email
addresses as a Set<String>. Testing in this set is faster than
looking in the cache, and yields simpler code.

Rename "get" to "list" as the property is more commonly called "list"
when its used to implement the list() method on the collection.

Replace Account.Id.get() == with .equals(). It is bad form to dig
into the contents of an object in order to test equality.

Change-Id: I75101a6a269cb679828290ce0dd833ec32a3f725

Former-commit-id: f6bd808ab578491a32485bdf2d3bd4bc4bd3273c