Add merge validation interface
Add a new MergeValidationListener interface which can be used to
provide additional validation of commits before they are merged to
the git repository.

Add a MergeValidators class to invoke the merge validator listeners,
and call it from MergeOp before the commit is merged.

Add a listener to invoke validation provided by plugins.

Change-Id: I325d923f5cc0245b60e86a035329b640c1682d48

Former-commit-id: 8f9e9f58e2c3ad65c30fecbc136b375e7979e7b1