Support adding an SSH key via REST
By POST on /accounts/<account-id>/sshkeys it is now possible to add
an SSH key for a user. The SSH public key has to be provided as raw
content in the request body.

The WebUI is adapted to use the new REST endpint to add SSH keys. The
old RPC for this is deleted.

Change-Id: I9617ccf19f078f9bcee615803755f445ed68bdc9
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 4e2e5216f9223f763b41440cf7de655d8e1d830d