Add an option to selectively disable changeId
The commit hook, if added, always adds a changeId to the bottom
of the commit message.

There is a configuration variable gerrit.createChangeId, which
can be used to configure EGit to generate the changeId if specified.

Respect the setting gerrit.createChangeId=false if the commit hook
is set, in case standard tools always put the commit hook in place
but it is not required for certain builds. This can be set at a
global, or local change.

Change-Id: I43a098453005fe0dcc60b6c4f6e4817f97ce24cf

Former-commit-id: b6131581f1f1f07f630f02add2b49c8444d2594d