Always set change status for ChangeAttributes
ChangeAttribute's status was only set upon extension of the
ChangeAttribute. As extension is only used for queries, the event
stream does not provide the change status. We now set the change
status when transforming a change to a ChangeAttribute. Thereby,
event stream consumers can for example detect whether or not an event
refers to a draft change, or a non-draft change.

Change-Id: I252eb5d757ab3575def7c75a033dbc5b9b6a2f4a

Former-commit-id: 4c1a37d37a9f5cad836b6851910d81216b1dfc80