BUCK: fix linking with Lucene
Make gerrit-lucene depend on reviewdb-server instead of -client.
This enables artifacts which depend on gerrit-lucene and on
reviewdb-server (gerrit-pgm) to be combined together into a
java_binary(). Otherwise it would not be possible because of
duplicate classes: reviewdb-server.jar contains classes in common
with reviewdb-client.jar.

The same holds true for LICENSE.txt and NOTICE.txt:
skip these files from Maven artifacts lucene-core.jar and
lucene-analyzers-common.jar. Otherwise it wouldn't be possible to
put artifacts which depend on gerrit-lucene into a java_binary().

Change-Id: I20a7f087b517a6ddadf58e88890fe68505177d48

Former-commit-id: 020d76b7f99a343e3e48385195e146f31a86ca62