Prefix start replication capability with plugin name
If startReplication global capability is used it must be migrated to
plugin-owned capability. By design the plugin-owned capabilities are prefixed
with the "plugin-name-" prefix, to ensure the uniqueness of the capability-id
across all plugins in `project.config` file. Given that the name of
replication-plugin is "replication.jar", the id of the "startReplication"
capability has to be changed to "replication-startReplication".

Increase schema version (81) and guess the name of replication plugin.
Because pgm-daemon not aware of PluginLoader, scan the plugin directory and
guess the name of replication plugin. If replication plugin is found, then its
name is used to prefix the existing "startReplication"-capability id. If it is
not found, then "startReplication"-capability is removed (when present) from the
`project.config`.

Change-Id: Id35ac8fca72946c9be61bdde129b8eaf8ed022af

Former-commit-id: 4e94ec20dad200e07fb09e5427b58bb1f615a647