Log all exceptions that occur while replacing a change in the index
Replacing a change in the Lucene index is done in the background.
Currently only IOExceptions are caught and logged in the Runnable.
RuntimeExceptions are silently ignored, which makes it hard to see
if and why indexing failed, e.g. reindexing changes from the classes
that implement the REST API is currently failing with a
ProvisionException due to a missing requestScopePropagator and this
exception is not visible from the logs.

Change-Id: Ia4ed727a1c38eded514992cbbf739ad52ac10d41
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 3c4943675daa6c6c0bbc5623116088d2848e7f36