Fix callback group implementation to match the documentation.
The contract for a CallbackGroup is that all callbacks are invoked
exactly once (either success or failure).
  + Failures are invoked immediately
  + Success wait until all callbacks are complete

In certain situations, like caches, callbacks may be invoked immediately
if the result is known. Depending on how the other callbacks completed,
this could have caused the callback to be invoked more than once; once for
success and again for the subsequently added callback success or failure.

Change-Id: If4c5d99bff6dc916826572ddb6fdff914536a01d

Former-commit-id: 332e67cd0578937c400dd4a7494f8ea4a54bf968