Fix ChangeIndexer to be runnable from REST API
The request scope should not be relied upon to obtain a database
connection.  The ChangeIndexer does not need the CurrentUser, it just
expects a Provider<ReviewDb> to work.  Temporarily obtain a ReviewDb
when the task starts executing, pass it through as part of the
RequestContext, and release the connection when the task is done.

Change-Id: I3ebcb634bfb7112756cdb2ad49f3ead83cd1b4d7

Former-commit-id: a5ca2095e494d026be964a4b565fb5311e92ddd6