Serve label predicate from Lucene index
To build the Lucene query we need to know the maximum range a label can
have. Since the label ranges can be defined per project the maximum
range for a label is hard to compute. For this we would need to scan
over all projects and parse the project.config files. Instead of doing
this we hard code for now a maximum query range from -4 to +4. This
range should be wide enough to support most labels.

The query behaviour for searching labels with value 0 has changed. We
now return all changes that have no voting on that label from any
reviewer (this includes changes without any reviewer). Earlier we
returned changes that had a reviewer for which an explicit 0 value was
stored. This old behaviour was unexpected by most users.

Change-Id: Ic0e5ae86933d605e67c9da8ef289a0b099d82cbf
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 451cda472b8ba0d19ae33bb665ab0fc32bd0907e