Fix copying of sticky approvals on patch set creation
The copying of the sticky approvals from the current patch set to the
newly created patch set in PatchSetInserter is currently failing since
the wrong patch set ID is provided as target patch set. It currently
tries to copy the sticky approvals from the current patch set to the
current patch set which fails with a primary key violation on insert.

The error e.g. happens when editing the commit message of a change that
has sticky approvals from the WebUI.

This bug is bad as it allows to remove a -2 veto voting from a change.

Change-Id: I443630f58823968bb6080a012d2df08b70b557b2
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: d365347cc59639e141dd0452cf2f71a1e0441fb6