Replace Reindex --dry-run with --output
The line of what was a "dry run" was a bit arbitrary, doing all the
Gerrit processing (including potential side effects like writing to
persistent caches) but not all the index-specific processing, which
could still be useful for evaluation or debugging purposes. Instead,
provide an output base so administrators can test the indexing
behavior without operating on production index data.

For Lucene, this flag is interpreted as a path to the index
directory; for Solr, it is a prefix added to the collection names.

Change-Id: Ic4808d1e3467a780a818ac7d1c89de13610da630

Former-commit-id: 97a01dbb05aeeffd2e88346528a4e99fd63f7b89