Make gwtexpui SafeHtml compatible with GWT SafeHtml
The GWT SafeHtml class is a simple interface that is already
compatible with gwtexpui SafeHtml. Implement the interface so
gwtexpui SafeHtml classes can be passed around to any GWT code.

The gwtexpui SafeHtmlBuilder is more flexible than the GWT one,
so it is still advantageous to keep gwtexpui SafeHtml in tree.

Change-Id: Ie985441dfe1bdddd0de7898ccb5db1aad2d8e972

Former-commit-id: 9023701849a3a46c464d5d93d89da8aad9f6c04a