Cache Not Found responses for retrieving avatar images
If an avatar image is not found it is likely that a moment later it is
still not found. Allowing to cache the Not Found response saves
requests to the server. If an avatar image gets available it is okay to
have some delay until it is refetched from the server.

Change-Id: I24ed6be3ea730fc604be036542638cad5118d6d4
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 95b4b20b71a248ab809db65d8d161ffb8aa863e7