Use legacy message:predicate when secondary index disabled
This commit is a partially revert of converting message:predicate to secondary
index based, falling back to db based message:predicate when secondary index is
disabled.

This provide an upgrade path to Gerrit users, who are not able to immediately
activate secondary index support and still want to have the whole feature set
after upgrading to the new version.

Change-Id: Ie5039218ecf05a078c770959802ac8a0300da8df

Former-commit-id: 513957f67a71d4fdf56e986abdf290ac86519b86