Allow plugins to provide an 'about' section on their documentation index
If a plugin provides a documentation file named "about.md|html", its
content will be inserted at the top of the index page above the lists
of documentation and commands.

If both "about.md" and "about.html" exist, only the first one to be
discovered from the JAR file will be used.

Change-Id: Id0de957b4cc5c2763e945310e252b59e27756639

Former-commit-id: 6853b5a996ecd93f2b30a10741f9ccd9fe6ba9a8