Disable caching for revision URLs using "current"
The current revision is resolved dynamically and may change. Do
not allow browser level caching on these URLs as the contents can
(and usually does) change in the future.

Clients that want caching must request the revision-id using a stable
identifier, ideally the commit SHA-1, or at least the unique patch
set number. This allows contents to cache as data behind the URL
will not change.

Change-Id: I00c5c607a78297dadaf5d5338f025aaf4832af63

Former-commit-id: c2cb3cda1dccfc8d8ceb904f202f1ee494570ac7