Filter out core UiCommands from legacy PatchSet view
Don't send any core commands to the legacy patch set view. These are
currently handled in a different API and UI display code, most of
which I am trying to rewrite and delete. Assume for now that there
will be no new interesting core commands while we switch the view.

Change-Id: If129a413d673ac1eb20981f0e25a0c9382e71ff9

Former-commit-id: 7e90035c0927c6d0dace891b617abf1dd83b3a32