Fix "Access-Control-Allow-Origin"-error on Safari browser
in Safari browser the site stays empty.

for JSONP request, jsonp=js_function must be passed,
or when callback=? passed, then jQuery replace it with temp method name.

jQuery API:

If the URL includes the string "callback=?" (or similar, as defined by the
server-side API), the request is treated as JSONP instead.

Change-Id: Ia19bb2bd0b7e6a391c5f8c349920189e7671a1d9

Former-commit-id: e19513ce997de720f3eb4ff28086b0d1e90d5bcd